Changes after first review
