+++ src/base-config/keyboard.json
@@ -221,24 +221,6 @@
             "displayKey": "Alt-↓"
         }
     ],
-    "debug.showDeveloperTools":  [
-        {
-            "key": "F12"
-        },
-        {
-            "key": "Cmd-Opt-I",
-            "platform": "mac"
-        }
-    ],
-    "debug.refreshWindow":  [
-        {
-            "key": "F5"
-        },
-        {
-            "key": "Cmd-R",
-            "platform": "mac"
-        }
-    ],
     "file.rename":  [
         "F2"
     ]
