+++ src/document/DocumentCommandHandlers.js
@@ -23,7 +23,7 @@
 
 
 /*jslint vars: true, plusplus: true, devel: true, nomen: true, indent: 4, maxerr: 50 */
-/*global define, $, brackets, PathUtils, WebSocket, window */
+/*global define, $, brackets, PathUtils, window */
 
 define(function (require, exports, module) {
     "use strict";
@@ -45,7 +45,6 @@ define(function (require, exports, module) {
         Async               = require("utils/Async"),
         Dialogs             = require("widgets/Dialogs"),
         Strings             = require("strings"),
-        PreferencesManager  = require("preferences/PreferencesManager"),
         PerfUtils           = require("utils/PerfUtils"),
         KeyEvent            = require("utils/KeyEvent");
     
@@ -508,43 +507,6 @@ define(function (require, exports, module) {
         return result.promise();
     }
     
-
-    /**
-     * Disables Brackets' cache via the remote debugging protocol.
-     * @return {$.Promise} A jQuery promise that will be resolved when the cache is disabled and be rejected in any other case
-     */
-    function _disableCache() {
-        var result = new $.Deferred();
-        
-        if (brackets.inBrowser) {
-            result.resolve();
-        } else {
-            var Inspector = require("LiveDevelopment/Inspector/Inspector");
-            Inspector.getAvailableSockets("127.0.0.1", "9234")
-                .fail(result.reject)
-                .done(function (response) {
-                    var page = response[0];
-                    if (!page || !page.webSocketDebuggerUrl) {
-                        result.reject();
-                        return;
-                    }
-                    var _socket = new WebSocket(page.webSocketDebuggerUrl);
-                    // Disable the cache
-                    _socket.onopen = function _onConnect() {
-                        _socket.send(JSON.stringify({ id: 1, method: "Network.setCacheDisabled", params: { "cacheDisabled": true } }));
-                    };
-                    // The first message will be the confirmation => disconnected to allow remote debugging of Brackets
-                    _socket.onmessage = function _onMessage(e) {
-                        _socket.close();
-                        result.resolve();
-                    };
-                    // In case of an error
-                    _socket.onerror = result.reject;
-                });
-        }
-            
-        return result.promise();
-    }
     
     /**
      * Closes the specified file: removes it from the working set, and closes the main editor if one
@@ -720,57 +682,17 @@ define(function (require, exports, module) {
         return result.promise();
     }
     
-    /**
-    * @private - tracks our closing state if we get called again
-    */
-    var _windowGoingAway = false;
-    
-    /**
-    * @private
-    * Common implementation for close/quit/reload which all mostly
-    * the same except for the final step
-    */
-    function _handleWindowGoingAway(commandData, postCloseHandler, failHandler) {
-        if (_windowGoingAway) {
-            //if we get called back while we're closing, then just return
-            return (new $.Deferred()).reject().promise();
-        }
-
-        return CommandManager.execute(Commands.FILE_CLOSE_ALL, { promptOnly: true })
-            .done(function () {
-                _windowGoingAway = true;
-                
-                // Give everyone a chance to save their state - but don't let any problems block
-                // us from quitting
-                try {
-                    $(ProjectManager).triggerHandler("beforeAppClose");
-                } catch (ex) {
-                    console.error(ex);
-                }
-                
-                PreferencesManager.savePreferences();
-                
-                postCloseHandler();
-            })
-            .fail(function () {
-                _windowGoingAway = false;
-                if (failHandler) {
-                    failHandler();
-                }
-            });
-    }
-
-    /**
-    * @private
-    * Implementation for abortQuit callback to reset quit sequence settings
-    */
+	/**
+     * @private
+     * Implementation for abortQuit callback to reset quit sequence settings
+     */
     function _handleAbortQuit() {
-        _windowGoingAway = false;
+        brackets.windowGoingAway = false;
     }
     
     /** Confirms any unsaved changes, then closes the window */
     function handleFileCloseWindow(commandData) {
-        return _handleWindowGoingAway(
+        return brackets.handleWindowGoingAway(
             commandData,
             function () {
                 window.close();
@@ -798,7 +720,7 @@ define(function (require, exports, module) {
 
     /** Closes the window, then quits the app */
     function handleFileQuit(commandData) {
-        return _handleWindowGoingAway(
+        return brackets.handleWindowGoingAway(
             commandData,
             function () {
                 brackets.app.quit();
@@ -813,16 +735,6 @@ define(function (require, exports, module) {
         );
     }
 
-    /** Does a full reload of the browser window */
-    function handleFileReload(commandData) {
-        return _handleWindowGoingAway(commandData, function () {
-            // Disable the cache to make reloads work
-            _disableCache().always(function () {
-                window.location.reload(true);
-            });
-        });
-    }
-    
     
     /** Are we already listening for a keyup to call detectDocumentNavEnd()? */
     var _addedNavKeyHandler = false;
@@ -909,7 +821,5 @@ define(function (require, exports, module) {
     // Listen for changes that require updating the editor titlebar
     $(DocumentManager).on("dirtyFlagChange", handleDirtyChange);
     $(DocumentManager).on("currentDocumentChange fileNameChange", updateDocumentTitle);
-    
-    exports.handleFileReload = handleFileReload;
-});
 
+});
