+++ src/extensions/default/DebugCommands/keyboard.json
@@ -1,0 +1,20 @@
+{
+    "showDeveloperTools":  [
+        {
+            "key": "F12"
+        },
+        {
+            "key": "Cmd-Opt-I",
+            "platform": "mac"
+        }
+    ],
+    "refreshWindow":  [
+        {
+            "key": "F5"
+        },
+        {
+            "key": "Cmd-R",
+            "platform": "mac"
+        }
+    ]
+}
